Paint the bounding rect of the dirty region instead of individual rectangles when the region covers at least 3/4 of its bounds.
Reviewed by Sam Weinig.

* WebProcess/WebPage/DrawingAreaImpl.cpp:
(WebKit::shouldPaintBoundsRect): Use floating point math so that the logic that was supposed to
achieve this actually works.



Canonical link: https://commits.webkit.org/74412@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84682 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/WebProcess/WebPage/DrawingAreaImpl.cpp
@@ -524,7 +524,7 @@ void DrawingAreaImpl::display()
 static bool shouldPaintBoundsRect(const IntRect& bounds, const Vector<IntRect>& rects)
 {
     const size_t rectThreshold = 10;
-    const float wastedSpaceThreshold = 0.75f;
+    const double wastedSpaceThreshold = 0.75;
 
     if (rects.size() <= 1 || rects.size() > rectThreshold)
         return true;
@@ -537,7 +537,7 @@ static bool shouldPaintBoundsRect(const IntRect& bounds, const Vector<IntRect>& 
     for (size_t i = 0; i < rects.size(); ++i)
         rectsArea += rects[i].width() * rects[i].height();
 
-    float wastedSpace = 1 - (rectsArea / boundsArea);
+    double wastedSpace = 1 - (static_cast<double>(rectsArea) / boundsArea);
 
     return wastedSpace <= wastedSpaceThreshold;
 }
